requirements.txt
pandas
numpy
pyyaml
python-dotenv
requests
lightgbm
xgboost
catboost
scikit-learn
mlflow
fastf1
lightgbm
joblib
matplotlib